api/requirements.txt
certifi==2020.12.5
chardet==4.0.0
clarifai-grpc==7.4.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
googleapis-common-protos==1.53.0
grpcio==1.38.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
PillowImage==1.1.11
protobuf==3.17.1
PyBundle==1.0.6
reportlab==3.5.67
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
pymongo==3.11.4
python-dotenv==0.17.1